added PGP keys for python releases on mac & windows
I just finished 3TOFU of these keys. Fortunately, python signs its releases with PGP, so once we track these keys, we can actually cryptographically verify the authenticity of future python releases for both platforms.

On Linux we'll continue to install from apt or use the AppImage.

On Windows it's the only option to install python.

On MacOS, we'll try using the python.org release instead of the brew release. Not only is this more secure (brew is insecure), but it's also more likely to be backwards compatible than the brew builds (to be tested still)

 * https://github.com/BusKill/buskill-app/issues/78#issuecomment-2028446273