improve set_trigger() fails
The last commit was successfully able to catch the failure to spawn the root_child process on setting the trigger to 'self-destruct' if the mode (permissions) on the root_child_process don't meet our security needs.

This commit improves the error message, and it also makes sure to revert-back the trigger to the old_trigger. Otherwise, it still tells the user that the trigger is 'soft-shutdown, even though that's impossible.

I probably should add something to the arm/disarm toggle that checks to see if the root_child is still running. something like a ping/pong "are you still alive?" test? If that fails, then we really should tell the user rather than letting them think that buskill is successfully armed.