add ping/pong to root_child_mac.py
This commit adds a new command named "ping" to the root child script for MacOS.

This command simply triggers the root_child to respond with "pong"

I updated the buskill class to attempt to "ping" the root_child script if it's attempting to arm buskill with 'soft-shutdown' on MacOS (currently the only platform + trigger combination that requires a root child process, because MacOS can't trigger a shutdown without root privileges)

I update the GUI to check for errors at arm-time (such as not being able to contact the root child process), and to *not* change the GUI from disarmed/blue to armed/red if it got an error. And it presents the error message (raised as an exception) to the user in a modal popup.

I haven't tested this yet.