attempt to catch exceptions from usb_handler child process
This commit adds a 'status' instance field to the multiprocessing.Process() child process, which starts-out at 0 and gets incremented by the child process at some unknown interval (actually its updated by the parent by listening to a 'ping' command placed into the queue by the child).

This allows the parent to check-on the state of the child process after launching it, which is important because exceptions thrown by the child won't raise an exception up to the parent (we could catch it, but that would be blocking).

This allows us to detect if there's issues immediately after the user clicks the "arm" button. After this commit, we can prevent the app from giving the user a false sense of security, because we'll throw an error on toggle() if the child isn't well and the GUI will remain blue/disarmed rather than saying it's armed (even if it can't possibly trigger because the child process has issues)

 * https://github.com/BusKill/buskill-app/issues/77#issuecomment-2253657854