s/trigger/buskill_trigger in Config
This commit replaces the Config item name "buskill" with "buskill_trigger"

Apparently I silently renamed this option in the Config file some months ago when I was switching the Settings screen to use RecycleView, for the font accessibility feature

 * https://github.com/BusKill/buskill-app/issues/55#issuecomment-2016994777

I don't know exactly why I made the change, but apparently I already updated it in other places, except here. The result: a bug where the first time you "arm" BusKill, it always defaulted to the 'lock-screen' trigger, instead of the setting the user already set it to.

 * https://github.com/BusKill/buskill-app/issues/77#issuecomment-2254691103